refactor: user /user/index.ts
