+++ packages/apps/reputation-oracle/server/src/modules/auth/auth.module.ts
@@ -4,7 +4,7 @@ import { JwtModule } from '@nestjs/jwt';
 import { AuthConfigService } from '../../config/auth-config.service';
 import { HCaptchaModule } from '../../integrations/hcaptcha/hcaptcha.module';
 import { EmailModule } from '../email/module';
-import { UserModule } from '../user/user.module';
+import { UserModule } from '../user';
 import { Web3Module } from '../web3/web3.module';
 
 import { JwtHttpStrategy } from './strategy';
