+++ packages/apps/reputation-oracle/server/src/modules/auth/token.entity.ts
@@ -6,7 +6,10 @@ import {
   JoinColumn,
   ManyToOne,
 } from 'typeorm';
-
+/**
+ * TODO: Leave fix follow-up refactoring
+ * Importing from '../user' causes circular import error here.
+ */
 import { UserEntity } from '../user/user.entity';
 import { BaseEntity } from '../../database/base.entity';
 import { DATABASE_SCHEMA_NAME } from '../../common/constants';
