+++ packages/apps/reputation-oracle/server/src/modules/kyc/kyc.module.ts
@@ -1,7 +1,7 @@
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
 
-import { UserModule } from '../user/user.module';
+import { UserModule } from '../user';
 import { Web3Module } from '../web3/web3.module';
 
 import { KycService } from './kyc.service';
