+++ packages/apps/reputation-oracle/server/src/modules/qualification/qualification.repository.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { BaseRepository } from '../../database/base.repository';
 import { DataSource, In, IsNull, MoreThan } from 'typeorm';
 import { QualificationEntity } from './qualification.entity';
-import { UserEntity } from '../user/user.entity';
+import { UserEntity } from '../user';
 import { UserQualificationEntity } from './user-qualification.entity';
 
 @Injectable()
