+++ packages/apps/reputation-oracle/server/src/modules/qualification/user-qualification.entity.ts
@@ -1,6 +1,10 @@
 import { Entity, ManyToOne } from 'typeorm';
 import { DATABASE_SCHEMA_NAME } from '../../common/constants';
 import { BaseEntity } from '../../database/base.entity';
+/**
+ * TODO: Leave fix follow-up refactoring
+ * Importing from '../user' causes circular import error here.
+ */
 import { UserEntity } from '../user/user.entity';
 import { QualificationEntity } from '../qualification/qualification.entity';
